Factor shader code into a separate file
